fix: Driver updater will now automatically restart
When driver updates are complete the updater will automatically restart unless any key is pressed. This was added in case the remote connection is severed due to a network driver update. The restart should auto reconnect the unit.

Additionally the software installers have been updated with the most up to date versions.